work on RA memory consumption
Most of the changes are in the respective RU and RP branches
(feature/prof_mem). We change from dict-based profile events to tuple
based ones, avoid duplicated storage, streamline input parsing, and
allow to filter out unnecessary events.  At the same time we move some
parts closer to the RU profile class, and only leave those parts in RP
which are dependent on the RP state model, and in RA which are unrelated
to profile reading and parsing.

NOTE: some of the legacy support for the old RP profile and DB format
      has been removed, which isgnificantly reduces the code complexity.
      This code should be used with RU/RP versions 0.46 or later.